Fix routes so we can have multiple components on the same page using the same ID. small designadjustment to battery to make it easier for the user to see their input. added name to avatar component as those two are tightly connected. restructured avatar functions and image array in a utils file to keep the user model clean. Set up the dash so its just to build the remaining components to make them appear.
