+++ backend/models/user-model.js
@@ -1,22 +1,8 @@
 import mongoose from 'mongoose';
-
 import bcryptjs from 'bcryptjs';
+import { getRandomAvatarUrl } from '../utils/avatarUtils.js';
 
 const SALT_ROUNDS = 12; // make this configurable so we can adjust the security level if needed
-
-const defaultAvatars = [
-  //An array of default avatar URLs from cloudinary
-  "https://res.cloudinary.com/egocloud/image/upload/v1717352040/ego-tree-avatar01.png",
-  "https://res.cloudinary.com/egocloud/image/upload/v1717352124/ego-tree-avatar02.png",
-  "https://res.cloudinary.com/egocloud/image/upload/v1717352124/ego-tree-avatar03.png",
-  "https://res.cloudinary.com/egocloud/image/upload/v1717352125/ego-tree-avatar04.png",
-];
-
-const getRandomAvatarUrl = () => {
-  const randomIndex = Math.floor(Math.random() * defaultAvatars.length);
-  return defaultAvatars[randomIndex];
-};
-
 
 const userSchema = new mongoose.Schema({
   username: {
