+++ frontend/src/components/Sidebar.jsx
@@ -2,12 +2,11 @@ import { Logo } from "../reusables/Logo";
 import { Avatar } from "../reusables/Avatar";
 import { Logout } from "../pages/Logout";
 
-export const Sidebar = ({id}) => {
+export const Sidebar = ({ id }) => {
   return (
     <>
       <Logo />
-      <Avatar id={id}/>
-      <p>Username</p>
+      <Avatar id={id} />
       <Logout />
     </>
   )
