+++ frontend/src/pages/Dashboard.jsx
@@ -9,11 +9,11 @@ export const Dashboard = () => {
   const { id } = useParams();
   return (
     <>
-      <Sidebar  id={id}/>
+      <Sidebar id={id} />
       <PointsCard id={id} />
       <ActivityGraph id={id} />
-      <DistanceCard  id={id} />
-      <EnergyCard id={id}/>
+      <DistanceCard id={id} />
+      <EnergyCard id={id} />
     </>
   )
 }
